소스코드/shoefinder/requirements.txt
absl_py==0.6.1
six==1.11.0
numpy==1.15.4
tensorflow==1.12.0
requests==2.18.4
Werkzeug==0.14.1
Flask==1.0.2
Pillow==5.3.0
google-api-python-client>=1.6.7
google-cloud-bigquery>=0.31.0
kaggle>=1.3.9
mlperf_compliance==0.0.10
oauth2client>=4.1.2
pandas
psutil>=5.4.3
py-cpuinfo>=3.3.0
Jinja2==2.10
urllib3==1.22
typing
